requirements.txt
imageio==2.5.0
numpy==1.16.2
Pillow==6.0.0
rawpy==0.13.1
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
